+++ actix-limitation/src/builder.rs
@@ -7,10 +7,19 @@ use redis::Client;
 
 use crate::{errors::Error, GetArcBoxKeyFn, Limiter};
 
+/// [RedisConnectionKind] is used to define which connection parameter for the Redis server will be passed
+/// It can be an Url or a Client
+/// This is done so we can use the same client for multiple Limiters
+#[derive(Debug, Clone)]
+pub enum RedisConnectionKind {
+    Url(String),
+    Client(Client),
+}
+
 /// Rate limiter builder.
 #[derive(Debug)]
 pub struct Builder {
-    pub(crate) redis_url: String,
+    pub(crate) redis_connection: RedisConnectionKind,
     pub(crate) limit: usize,
     pub(crate) period: Duration,
     pub(crate) get_key_fn: Option<GetArcBoxKeyFn>,
@@ -96,8 +105,12 @@ impl Builder {
             closure
         };
 
+        let client = match &self.redis_connection {
+            RedisConnectionKind::Url(url) => Client::open(url.as_str())?,
+            RedisConnectionKind::Client(client) => client.clone(),
+        };
         Ok(Limiter {
-            client: Client::open(self.redis_url.as_str())?,
+            client,
             limit: self.limit,
             period: self.period,
             get_key_fn: get_key,
@@ -109,12 +122,25 @@ impl Builder {
 mod tests {
     use super::*;
 
+    /// Implementing partial Eq to check if builder assigned the redis connection url correctly
+    /// We can't / shouldn't compare Redis clients, thus the method panic if we try to compare them
+    impl PartialEq for RedisConnectionKind {
+        fn eq(&self, other: &Self) -> bool {
+            match (self, other) {
+                (Self::Url(l_url), Self::Url(r_url)) => l_url == r_url,
+                _ => {
+                    panic!("RedisConnectionKind PartialEq is only implemented for Url")
+                }
+            }
+        }
+    }
+
     #[test]
     fn test_create_builder() {
-        let redis_url = "redis://127.0.0.1";
+        let redis_connection = RedisConnectionKind::Url("redis://127.0.0.1".to_string());
         let period = Duration::from_secs(10);
         let builder = Builder {
-            redis_url: redis_url.to_owned(),
+            redis_connection: redis_connection.clone(),
             limit: 100,
             period,
             get_key_fn: Some(Arc::new(|_| None)),
@@ -123,7 +149,7 @@ mod tests {
             session_key: Cow::Owned("rate-api".to_string()),
         };
 
-        assert_eq!(builder.redis_url, redis_url);
+        assert_eq!(builder.redis_connection, redis_connection);
         assert_eq!(builder.limit, 100);
         assert_eq!(builder.period, period);
         #[cfg(feature = "session")]
@@ -133,10 +159,10 @@ mod tests {
 
     #[test]
     fn test_create_limiter() {
-        let redis_url = "redis://127.0.0.1";
+        let redis_connection = RedisConnectionKind::Url("redis://127.0.0.1".to_string());
         let period = Duration::from_secs(20);
         let mut builder = Builder {
-            redis_url: redis_url.to_owned(),
+            redis_connection,
             limit: 100,
             period: Duration::from_secs(10),
             get_key_fn: Some(Arc::new(|_| None)),
@@ -154,10 +180,10 @@ mod tests {
     #[test]
     #[should_panic = "Redis URL did not parse"]
     fn test_create_limiter_error() {
-        let redis_url = "127.0.0.1";
+        let redis_connection = RedisConnectionKind::Url("127.0.0.1".to_string());
         let period = Duration::from_secs(20);
         let mut builder = Builder {
-            redis_url: redis_url.to_owned(),
+            redis_connection,
             limit: 100,
             period: Duration::from_secs(10),
             get_key_fn: Some(Arc::new(|_| None)),