+++ actix-limitation/CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
-## Unreleased
+## Unreleased - 2023-xx-xx
+
+- Added optional scopes to the middleware enabling use of multiple Limiters by passing an `HashMap<Limiter>` to the Http server `app_data`
 
 - Update `redis` dependency to `0.32`.
 - Update `actix-session` dependency to `0.11`.